+++ src/commands/Main/task.js
@@ -1,3 +1,6 @@
+const { MessageEmbed } = require("discord.js");
+const config = require("../../../configs/token.json");
+const resource = require("../../../configs/resource.json");
 module.exports = {
   config: {
     name: `task`,
