+++ src/commands/Social/docs.js
@@ -1,4 +1,6 @@
 const Discord = require('discord.js');
+const config = require("../../../configs/token.json");
+const resource = require("../../../configs/resource.json");
 module.exports = {
   config: {
     name: `userdocs`,
@@ -20,7 +22,7 @@ module.exports = {
         }, 1000);
       });
     } catch (err) {
-      const embed = new MessageEmbed()
+      const embed = new Discord.MessageEmbed()
         .setTitle("Whoops, looks like something went wrong!")
         .setThumbnail(resource.aw_snap)
         .setDescription(
