+++ src/commands/Social/register_self.js
@@ -2,6 +2,7 @@
 const { Database } = require('secure-db');
 const Discord = require("discord.js");
 const config = require("../../../configs/token.json");
+const resource = require("../../../configs/resource.json");
 const moment = require("moment");
 module.exports = {
   config: {
@@ -59,7 +60,7 @@ module.exports = {
         }, 1000);
       });
     } catch (err) {
-      const embed = new MessageEmbed()
+      const embed = new Discord.MessageEmbed()
         .setTitle("Whoops, looks like something went wrong!")
         .setThumbnail(resource.aw_snap)
         .setDescription(
