+++ src/commands/science/Astronomy/iss.js
@@ -1,6 +1,5 @@
 const { MessageEmbed } = require("discord.js");
 const config = require("../../../../configs/token.json");
-const chnl = require("../../../../configs/chnl.json");
 const resource = require("../../../../configs/resource.json");
 const superagent = require("superagent");
 module.exports = {
