+++ src/commands/science/Astronomy/moonPhases.js
@@ -1,9 +1,6 @@
 const { MessageEmbed } = require("discord.js");
 const config = require("../../../../configs/token.json");
-const chnl = require("../../../../configs/chnl.json");
 const resource = require("../../../../configs/resource.json");
-const superagent = require("superagent");
-const colors = require("../../../../configs/colors.json");
 const Moon = require("lunarphase-js");
 const mphase = require("../../../../configs/mphase.json");
 module.exports = {
