+++ src/commands/science/Chemistry/historyOfElement.js
@@ -2,7 +2,6 @@ const { MessageEmbed } = require("discord.js");
 const config = require("../../../../configs/token.json");
 const chnl = require("../../../../configs/chnl.json");
 const resource = require("../../../../configs/resource.json");
-const superagent = require("superagent");
 const colors = require("../../../../configs/colors.json");
 var elementHistories = require("element-histories");
 module.exports = {
