+++ src/commands/science/Chemistry/molar.js
@@ -1,8 +1,6 @@
 const { MessageEmbed } = require("discord.js");
 const config = require("../../../../configs/token.json");
-const chnl = require("../../../../configs/chnl.json");
 const resource = require("../../../../configs/resource.json");
-const superagent = require("superagent");
 const colors = require("../../../../configs/colors.json");
 const molarcalc = require("molarcalc");
 module.exports = {
